pyproject.toml
[project]
name = "ai-art-shield"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.1",
    "pillow>=11.2.1",
    "pytest>=8.4.1",
    "pytest-mock>=3.14.1",
    "scipy>=1.16.0",
]
